requirements.txt
beautifulsoup4==4.12.0
cryptography==39.0.2
discord.py==2.2.2
fake_useragent==1.1.3
fastapi==0.104.1
getch==1.0
Pillow==9.4.0
Pillow==10.0.1
Pillow==10.1.0
psutil==5.9.5
pyradios==2.0.0
python_gnupg==0.5.1
Requests==2.31.0
uvicorn==0.23.2
yt_dlp==2023.10.13
yt_dlp==2023.9.24
